refactor: Service 다수 메서드 집합으로 분리
