+++ src/main/java/order/controller/OrderController.java
@@ -2,6 +2,7 @@ package order.controller;
 
 import order.dto.OrderRequestDTO;
 import order.dto.OrderResponseDTO;
+import order.service.OrderFactory;
 import order.service.OrderService;
 import order.util.ProductConverter;
 import order.validation.OrderValidator;
@@ -14,10 +15,13 @@ import java.util.List;
 public class OrderController {
     OutputView outputView = new OutputView();
     InputView inputView = new InputView();
+    OrderView orderView = new OrderView(outputView);
+
     ProductConverter productConverter = new ProductConverter();
     OrderValidator orderValidator = new OrderValidator();
-    OrderService orderService = new OrderService(orderValidator);
-    OrderView orderView = new OrderView(outputView);
+
+    OrderFactory orderFactory = new OrderFactory();
+    OrderService orderService = new OrderService(orderValidator, orderFactory);
 
     public void run() {
         orderView.orderStart();
